Move Opus encoding/decoding classes to own file
