+++ Sources/SwiftDiscord/Voice/DiscordVoiceDataSource.swift
@@ -377,80 +377,3 @@ public class DiscordEncoderMiddleware {
     }
 }
 #endif
-
-///
-/// An Opus encoder.
-///
-/// Takes raw PCM 16-bit-lesample data and returns Opus encoded voice packets.
-///
-open class DiscordOpusEncoder : DiscordOpusCodeable {
-    // MARK: Properties
-
-    /// The bitrate for this encoder.
-    public let bitrate: Int
-
-    /// The number of channels.
-    public let channels: Int
-
-    /// The maximum size of a opus packet.
-    public var maxPacketSize: Int { return 4000 }
-
-    /// The sampling rate.
-    public let sampleRate: Int
-
-    private let encoderState: OpaquePointer
-
-    // MARK: Initializers
-
-    ///
-    /// Creates an Encoder that can take raw PCM data and create Opus packets.
-    ///
-    /// - parameter bitrate: The target bitrate for this encoder.
-    /// - parameter sampleRate: The sample rate for the encoder. Discord expects this to be 48k.
-    /// - parameter channels: The number of channels in the stream to encode, should always be 2.
-    ///
-    public init(bitrate: Int, sampleRate: Int, channels: Int, vbr: Bool = false) throws {
-        self.bitrate = bitrate
-        self.sampleRate = sampleRate
-        self.channels = channels
-
-        var err = 0 as Int32
-
-        encoderState = opus_encoder_create(Int32(sampleRate), Int32(channels), OPUS_APPLICATION_VOIP, &err)
-        err = configure_encoder(encoderState, Int32(bitrate), vbr ? 1 : 0)
-
-        guard err == 0 else {
-            destroyState()
-
-            throw DiscordVoiceError.creationFail
-        }
-    }
-
-    deinit {
-        destroyState()
-    }
-
-    // MARK: Methods
-
-    private func destroyState() {
-        opus_encoder_destroy(encoderState)
-    }
-
-    ///
-    /// Encodes a single frame of raw PCM 16-bit-le/sample LE data into Opus format.
-    ///
-    /// - parameter audio: A pointer to the audio data. Use `maxFrameSize(assumingSize:)` to find the length.
-    /// - parameter frameSize: The size of the frame in samples per channel.
-    /// - returns: An opus encoded packet.
-    ///
-    open func encode(_ audio: UnsafePointer<opus_int16>, frameSize: Int) throws -> [UInt8] {
-        let output = UnsafeMutablePointer<UInt8>.allocate(capacity: maxPacketSize)
-        let lenPacket = opus_encode(encoderState, audio, Int32(frameSize), output, opus_int32(maxPacketSize))
-
-        defer { free(output) }
-
-        guard lenPacket > 0 else { throw DiscordVoiceError.encodeFail }
-
-        return Array(UnsafeBufferPointer(start: output, count: Int(lenPacket)))
-    }
-}
