+++ Sources/SwiftDiscord/Voice/DiscordVoiceEngineSpec.swift
@@ -131,41 +131,6 @@ public enum DiscordVoiceError : Error {
     case initialPacket
 }
 
-/// Declares that a type has enough information to encode/decode Opus data.
-public protocol DiscordOpusCodeable {
-    // MARK: Properties
-
-    /// The number of channels.
-    var channels: Int { get }
-
-    /// The sampling rate.
-    var sampleRate: Int { get }
-
-    // MARK: Methods
-
-    ///
-    /// Returns the maximum number of bytes that a frame can contain given a
-    /// frame size in number of samples per channel.
-    ///
-    /// - parameter assumingSize: The size of the frame, in number of samples per channel.
-    /// - returns: The number of bytes in this frame.
-    ///
-    func maxFrameSize(assumingSize size: Int) -> Int
-}
-
-public extension DiscordOpusCodeable {
-    ///
-    /// Returns the maximum number of bytes that a frame can contain given a
-    /// frame size in number of samples per channel.
-    ///
-    /// - parameter assumingSize: The size of the frame, in number of samples per channel.
-    /// - returns: The number of bytes in this frame.
-    ///
-    public func maxFrameSize(assumingSize size: Int) -> Int {
-        return size * channels * MemoryLayout<opus_int16>.size
-    }
-}
-
 /// A struct that is used to configure the high-level functions of a VoiceEngine
 public struct DiscordVoiceEngineConfiguration {
     /// Whether or not this engine should capture voice.
